python-sidecar/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
sherpa-onnx>=1.10.0
llama-cpp-python>=0.2.0
numpy>=1.24.0
scipy>=1.11.0
